hw/pci-host/designware: Support 64Bit resource for DesignwarePCIEHost
0. DW pcie support 64bit resource
1. DW version type using Linux kernel

Message-Id: <20250715073251.2376473-1-zhangzhijie@bosc.ac.cn>

+++ include/hw/pci-host/designware.h
@@ -46,7 +46,7 @@ typedef struct DesignwarePCIEViewport {
 
     uint64_t base;
     uint64_t target;
-    uint32_t limit;
+    uint64_t limit;
     uint32_t cr[2];
 
     bool inbound;
